Change app.js to app.tsx and read user input and pass it in parse funcion.
